refactor(contracts): move skill_contract.py wholesale to pawai_contracts + compat shim — zero behavior change (Plan C2)
Hook smart-scope PYTHONPATH gains pawai_contracts: prefix for IE/brain
suites (the shim made the hook itself prove the wiring gap by blocking
this commit's first attempt).
